+++ src/data/options/styleOptions.js
@@ -191,7 +191,8 @@ const defaultCandle = {
         paddingBottom: 2,
         color: '#FFFFFF',
         family: 'Helvetica Neue',
-        weight: 'normal'
+        weight: 'normal',
+        borderRadius: 2
       }
     }
   },
@@ -261,7 +262,8 @@ const defaultTechnicalIndicator = {
       paddingLeft: 3,
       paddingTop: 2,
       paddingRight: 3,
-      paddingBottom: 2
+      paddingBottom: 2,
+      borderRadius: 2
     }
   },
   tooltip: {
@@ -404,6 +406,7 @@ const defaultCrosshair = {
       paddingBottom: 2,
       borderSize: 1,
       borderColor: '#505050',
+      borderRadius: 2,
       backgroundColor: '#505050'
     }
   },
@@ -427,6 +430,7 @@ const defaultCrosshair = {
       paddingTop: 2,
       paddingBottom: 2,
       borderSize: 1,
+      borderRadius: 2,
       borderColor: '#505050',
       backgroundColor: '#505050'
     }